chore: move bootstrap code to service layer
